app/requirements.txt
faker==13.3.4
pandas==1.4.2
kafka-python==2.0.2
